apply other code review suggestions
